Revert "Feature/header fix"
