+++ client/src/index.css
@@ -16,7 +16,7 @@ body {
 
 .bio-div {
   background-color: #fff;
-  margin: 0 10px;
+  margin: 0 200px;
   padding: 10px;
   border-radius: 0.5rem;
 }
@@ -38,10 +38,11 @@ body {
 }
 
 .edit-button-img {
-  padding: 8px 0 0px;
+  max-width: 30px;
+  padding: 0 0 8px;
 }
 .edit-button-img:active {
-  padding: 0 0 8px;
+  padding: 0px;
 }
 
 .edit-subj-img{
@@ -75,6 +76,7 @@ body {
 }
 #edit-profile-btn {
   margin: 25% 0;
+  height: 34px;
   color: #6B7280;
   box-shadow: 0px 3px 2px #46484D;
 }
@@ -137,15 +139,9 @@ body {
     display: none;
   }
 
-  #edit-profile-btn {
+  .edit-profile {
     display: none;
   }
-
-  #add-subject-btn {
-    margin: 10px;
-    align-self: center;
-  }
-
   .home-div {
     margin: 210px 0 0;
   }
